refactor attention out of transformer.py
